+++ cmd/xkcd/config.go
@@ -1,4 +1,4 @@
-package config
+package main
 
 import (
 	"flag"
