+++ internal/core/service/search.go
@@ -1,4 +1,4 @@
-package app
+package service
 
 import (
 	"slices"
