+++ internal/core/service/search_test.go
@@ -1,14 +1,14 @@
-package app
+package service
 
 import (
-	"github.com/AfoninaOlga/xkcd/pkg/database"
+	"github.com/AfoninaOlga/xkcd/internal/adapter/repository/json"
 	"testing"
 )
 
 var keywords = []string{"account", "zip", "zero", "know", "question", "complain", "overlap", "live", "guess", "truth", "save", "bug"}
 
 func BenchmarkDB(b *testing.B) {
-	db, _ := database.New("../../database.json")
+	db, _ := json.New("../../database.json")
 	app := App{db: db}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -17,7 +17,7 @@ func BenchmarkDB(b *testing.B) {
 }
 
 func BenchmarkIndex(b *testing.B) {
-	db, _ := database.New("../../database.json")
+	db, _ := json.New("../../database.json")
 	app := App{db: db}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
